Created bower file to manage dependencies, and created initial files.
Created bower.json file to manage dependencies. Since I don't want to push
the large bower_components folder to github I added the line
"bower_components/" to the gitignore. I also created the basic folder
structure for the project and created some basic files to fill them in.
index.html is the only file that has any kind of content inside, with
references to the dependencies and the project files.