Added locations, filtering, and markers.
The main MVVM has been added to the new app.js file. Inside the file is an
array holding basic locations. These locations are added to an observable
array to be filtered later. The data bindings for the location list have
been added to index.html. Functionality for clicking is halfway complete:
an observable holds the currently selected location but there is no
feedback on the ui for the selection. Filtering works, and will rerender
the list and markers depending on what filter option is selected. Markers
have been added with an addMarker function added to map.js. Markers are
created and stored within their corresponding Location object as a normal
value, no need to make it observable as it's never going to change, just
need to call methods on it occasionally.

The google maps API script tag has been changed to be synchronous as the
'async defer' option made things difficult because app.js relies on google
maps API code as well.