Implemented various improvements, such as asynchrony for the Map.
Refactored the code to wait for the Google Map to load Asynchronously. If
the map API fails to load, various points in the code will not run, but
the application can still serve a basic, map-less version. anim.js has
been removed as its functionality has been reimplemented in knockout. As
such, much of it is now in index.html and app.js. The readme has been made
slightly clearer.